+++ lib/pageflow/configuration.rb
@@ -60,6 +60,11 @@ module Pageflow
     # @return [Themes]
     attr_reader :themes
 
+    # Register new types of pages.
+    # @return [PageTypes]
+    # @since edge
+    attr_reader :page_types
+
     # List of {FileType} instances provided by page types.
     # @return [FileTypes]
     attr_reader :file_types
@@ -178,6 +183,7 @@ module Pageflow
       @hooks = Hooks.new
       @quotas = Quotas.new
       @themes = Themes.new
+      @page_types = PageTypes.new
       @file_types = FileTypes.new(page_types)
       @widget_types = WidgetTypes.new
       @help_entries = HelpEntries.new
@@ -204,24 +210,10 @@ module Pageflow
       plugin.configure(self)
     end
 
-    # Make a page type available for use in the system.
+    # @deprecated Use `config.page_types.register` instead.
     def register_page_type(page_type)
-      page_types << page_type
-
-      @page_types_by_name ||= {}
-      @page_types_by_name[page_type.name] = page_type
-    end
-
-    def lookup_page_type(name)
-      @page_types_by_name.fetch(name)
-    end
-
-    def page_types
-      @page_types ||= []
-    end
-
-    def page_type_names
-      page_types.map(&:name)
+      ActiveSupport::Deprecation.warn('Pageflow::Configuration#register_page_type is deprecated. Use config.page_types.register instead.', caller)
+      page_types.register(page_type)
     end
 
     def revision_components