+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 
 ### Changes on `master`
 
-None so far
-
 [Compare changes](https://github.com/codevise/pageflow/compare/v0.8.0...master)
+
+##### Deprecations
+
+- The configuration API for registering new page types has
+  changed. Turn calls of the form
+
+        config.register_page_type(Pageflow::Some.page_type)
+
+  into
+
+        config.page_types.register(Pageflow::Some.page_type)
+
+  The old method still exists for now, but will be removed before
+  version 1.0.
 
 ### Version 0.8.0